mocha@^6.1.4 (#6)
It would be nice to update all the packages in this repo. ~This is as far as I can move Mocha before encountering a breaking change.~ I was able to fix the `yarn test` command so that mocha could be upgraded all the way to the latest version. This should resolve at least two of the Security Alerts for this repo.

Since there isn't CI right now I manually ran
`yarn eslint` and `yarn test` and they both work.