feat(deps): bump eslint-config-airbnb from 18 to 19 (#43)
This version bump contains lots of good rules and opens the path to `eslint@8`.

Turning off the rule `react/function-component-definition` as there is no clear consensus on what we use. We *mostly* use arrow functions which is an option but not the default.

Additionally ZP had quite a few fixes to other rules required and needs a bunch of instances grandfathered, but they all seem good to adopt.